+++ CHANGELOG.md
@@ -27,6 +27,7 @@ API after
 ```python
 client = substra.Client(backend_type=substra.BackendType.LOCAL_SUBPROCESS)
 ```
+- Pass substra-tools arguments via a file instead of the command line
 
 ### Removed